refactor(intelligence): decompose TrainingIntelligence into Facade + 4 Mixins
Extract models (2 enums + 3 dataclasses) into models.py and split
912-line god class into CoreMixin, InsightsMixin, PIDMixin, StorageMixin.
Facade re-exports all symbols for backward compatibility (13 importers).
