+++ magma_cycling/intelligence/training/__init__.py
@@ -1,0 +1,1 @@
+"""TrainingIntelligence mixin modules for god class decomposition."""
